requirements.txt
streamlit
torch
opencv-python
Pillow
numpy
os
time
json
kafka.KafkaProducer
kafka.KafkaConsumer
pandas 
torchvision.transforms
torchvision.models
plotly.express